refactor: 나이별 할인 관련 로직 객체로 분리
