+++ src/test/java/wooteco/subway/domain/FareTest.java
@@ -36,18 +36,4 @@ class FareTest {
     void calculateFareWithExtraFare() {
         assertThat(Fare.calculate(10, 500)).isEqualTo(1750);
     }
-
-    @ParameterizedTest
-    @ValueSource(ints = {6, 12})
-    @DisplayName("나이가 6 세 이상 13 세 미만일 경우 350 원을 공제한 금액에서 50% 를 할인한다.")
-    void discountChild(int age) {
-        assertThat(Fare.discount(1350, age)).isEqualTo(500);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {13, 18})
-    @DisplayName("나이가 13 세 이상 19 세 미만일 경우 350 원을 공제한 금액에서 20% 를 할인한다.")
-    void discountTeenager(int age) {
-        assertThat(Fare.discount(1350, age)).isEqualTo(800);
-    }
 }
